requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
grpcio==1.36.0
grpcio-tools==1.36.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
protobuf==3.15.3
requests==2.25.1
six==1.15.0
urllib3==1.26.3
Werkzeug==1.0.1
pika==1.2.0